upgraded robot library

+++ MyoSpheroControl/src/de/nachregenkommtsonne/myospherocontrol/controller/sphero/ISpheroMovementController.java
@@ -4,7 +4,7 @@ public interface ISpheroMovementController
 {
 	public abstract void halt();
 
-	public abstract void changeColor(int red, int green, int blue);
+	public abstract void changeColor(float red, float green, float blue);
 
 	public abstract void move(float direction, float speed);
 }
