+++ MyoSpheroControl/src/de/nachregenkommtsonne/myospherocontrol/movement/MovementResult.java
@@ -5,11 +5,11 @@ public class MovementResult
   private float _direction;
   private float _speed;
   
-  private int _red;
-  private int _green;
-  private int _blue;
+  private float _red;
+  private float _green;
+  private float _blue;
 
-  public MovementResult(float direction, float speed, int red, int green, int blue)
+  public MovementResult(float direction, float speed, float red, float green, float blue)
   {
     super();
     
@@ -21,17 +21,17 @@ public class MovementResult
     _blue = blue;
   }
 
-  public int getRed()
+  public float getRed()
   {
     return _red;
   }
 
-  public int getGreen()
+  public float getGreen()
   {
     return _green;
   }
 
-  public int getBlue()
+  public float getBlue()
   {
     return _blue;
   }
